Dev 77 Added ViewModel

+++ app/src/main/java/ru/transport/threeka/api/schemas/Coord.kt
@@ -1,6 +1,6 @@
 package ru.transport.threeka.api.schemas
 
 data class Coord(
-    val lat: Float,
-    val lon: Float
+    val lat: Double,
+    val lon: Double
 )
